Event records save properly to core data. Can load in entire list of events, and to see them click edit button on Medical Event Table page

+++ Mobile Computing Project App/Mobile Computing Project App/PainLevelCell.swift
@@ -18,7 +18,7 @@ class PainLevelCell: UITableViewCell {
     var row:Int?
     
     @IBAction func sendPainLevelNotification(_ sender: Any) {
-        let roundedVal:Int = Int(roundf(self.painSlider.value))
+        let roundedVal:Int16 = Int16(roundf(self.painSlider.value))
         self.painLevel.text = String(roundedVal)
         self.painSlider.value = Float(roundedVal)
         let dataDict:Dictionary<String, Any> = ["data":roundedVal,
